requirements.txt
cython
cytoolz
eth-brownie

rich
isort
